+++ ContentPatcher/Framework/Migrations/Migration_1_30.cs
@@ -3,15 +3,15 @@ using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
 {
-    /// <summary>Migrates patches to format version 1.29.</summary>
+    /// <summary>Migrates patches to format version 1.30.</summary>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
-    internal class Migration_1_29 : BaseMigration
+    internal class Migration_1_30 : BaseMigration
     {
         /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
-        public Migration_1_29()
-            : base(new SemanticVersion(1, 29, 0)) { }
+        public Migration_1_30()
+            : base(new SemanticVersion(1, 30, 0)) { }
     }
 }